Solution refactor
