+++ src/TimeIt.Core/ConsoleExporter.cs
@@ -5,7 +5,7 @@ using TimeIt.Common.Exporters;
 using TimeIt.Common.Results;
 using Status = TimeIt.Common.Results.Status;
 
-namespace TimeIt;
+namespace TimeIt.Core;
 
 public sealed class ConsoleExporter : IExporter
 {
