+++ src/TimeIt.StartupHook/TimeIt.StartupHook.csproj
@@ -5,7 +5,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <Compile Include="..\TimeIt\Constants.cs">
+        <Compile Include="..\TimeIt.Core\Constants.cs">
             <Link>Constants.cs</Link>
         </Compile>
     </ItemGroup>
