+++ src/TimeIt/TimeIt.csproj
@@ -16,12 +16,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="MathNet.Numerics" Version="5.0.0" />
-        <PackageReference Include="Spectre.Console" Version="0.47.0" />
-        <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
-    </ItemGroup>
-
-    <ItemGroup>
         <None Update="config-example.json">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
@@ -30,9 +24,7 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\TimeIt.Common\TimeIt.Common.csproj" />
-        <ProjectReference Include="..\TimeIt.StartupHook\TimeIt.StartupHook.csproj" />
-        <ProjectReference Include="..\TimeIt.DatadogExporter\TimeIt.DatadogExporter.csproj" />
+        <ProjectReference Include="..\TimeIt.Core\TimeIt.Core.csproj" />
     </ItemGroup>
 
 </Project>
